The image is complete in principle but I still need to test it some more.
VS Code Docker vulnerability scanning is turned off for this image
 because I have mitigated critical (there were none), high, medium,
 and low vulnerabilities and it's going to continue bugging me about
 them until nginx fixes them in the base image.
This is implemented in the workspace-local settings.json in this repo.

I have implemented development and production Dockerfiles for this so
 during development we can take advantage of the build cache.
Basically, this image:
 - Has configuration for running a reverse proxy to the HSF website
 - Follows best-practices for SSL/TLS
 - Uses certbot for certificate issuance and renewal for the HSF
    domains (base and www).

The entrypoint script handles the certificate management side of
 things.

Updated the README with additional information.