Update docker-image.yml
Don't fail the build to ensure the SARIF results can be posted to the repository's CodeQL results section. Otherwise I don't know what the vulnerabilities are that caused the build to fail.
Docker Scout did not show any Critical/Grave vulnerabilities, so it is surprising to me that it is finding critical-or-worse vulnerabilities in the image.